bluebird shim is needed
browserify won't let me require bluebird unless I have the entry
point inside node_modules.  This is dumb, but I don't know of a
better workaround.  I asked about it on #browserify, but received
no answer, so I have resolved to add a "shim" for bluebird for the
time being.